requirements.txt
numpy
Pillow
lama-cleaner